+++ {{cookiecutter.buildfolder}}/utils/delete.py
@@ -47,11 +47,13 @@ def delete_raw(ToolGlobals: CDFToolConfig, dry_run=False) -> None:
         )
 
 
-def delete_files(ToolGlobals: CDFToolConfig, dry_run=False) -> None:
+def delete_files(ToolGlobals: CDFToolConfig, dry_run=False, directory=None) -> None:
+    if directory is None:
+        directory = f"./examples/{ToolGlobals.example}/data/files"
     client = ToolGlobals.verify_client(capabilities={"filesAcl": ["READ", "WRITE"]})
     files = []
     # Pick up all the files in the files folder of the example.
-    for _, _, filenames in os.walk(f"./examples/{ToolGlobals.example}/data/files"):
+    for _, _, filenames in os.walk(directory):
         for f in filenames:
             files.append(ToolGlobals.example + "_" + f)
     if len(files) == 0:
@@ -72,29 +74,32 @@ def delete_files(ToolGlobals: CDFToolConfig, dry_run=False) -> None:
     )
 
 
-def delete_timeseries(ToolGlobals: CDFToolConfig, dry_run=False) -> None:
+def delete_timeseries(
+    ToolGlobals: CDFToolConfig, dry_run=False, directory=None
+) -> None:
     """Delete timeseries from CDF based on json files in the example"""
 
+    if directory is None:
+        directory = f"./examples/{ToolGlobals.example}/data/timeseries"
     client = ToolGlobals.verify_client(
         capabilities={"timeseriesAcl": ["READ", "WRITE"]}
     )
     files = []
     # Pick up all the .json files in the data folder of the example.
-    for _, _, filenames in os.walk(
-        f"./examples/{ToolGlobals.example}/data/timeseries/"
-    ):
+    for _, _, filenames in os.walk(directory):
         for f in filenames:
             if ".json" in f:
                 files.append(f)
     # Read timeseries metadata to build a list of TimeSeries
     timeseries: list[TimeSeries] = []
     for f in files:
-        with open(
-            f"./examples/{ToolGlobals.example}/data/timeseries/{f}", "rt"
-        ) as file:
+        with open(f"{directory}/{f}", "rt") as file:
             ts = json.load(file)
             for t in ts:
-                timeseries.append(TimeSeries._load(t))
+                ts = TimeSeries()
+                for k, v in t.items():
+                    ts.__setattr__(k, v)
+                timeseries.append(ts)
     if len(timeseries) == 0:
         return
     drop_ts: list[str] = []
@@ -116,13 +121,15 @@ def delete_timeseries(ToolGlobals: CDFToolConfig, dry_run=False) -> None:
         )
 
 
-def delete_transformations(ToolGlobals: CDFToolConfig, dry_run=False) -> None:
+def delete_transformations(
+    ToolGlobals: CDFToolConfig, dry_run=False, directory=None
+) -> None:
+    if directory is None:
+        directory = f"./examples/{ToolGlobals.example}/transformations"
     client = ToolGlobals.verify_client(
         capabilities={"transformationsAcl": ["READ", "WRITE"]}
     )
-    configs = parse_transformation_configs(
-        f"./examples/{ToolGlobals.example}/transformations/"
-    )
+    configs = parse_transformation_configs(directory)
     transformations_ext_ids = [t.external_id for t in configs.values()]
     try:
         if not dry_run: