+++ {{cookiecutter.buildfolder}}/utils/load.py
@@ -15,7 +15,9 @@
 import os
 import json
 import pandas as pd
+from typing import List, Dict, Any
 from cognite.client.data_classes.time_series import TimeSeries
+from cognite.client.data_classes.iam import Group
 from .utils import CDFToolConfig
 from utils.transformations_config import parse_transformation_configs
 from utils.transformations_api import (
@@ -27,13 +29,15 @@ from utils.transformations_api import (
 from cognite.client.exceptions import CogniteNotFoundError
 
 
-def load_raw(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
+def load_raw(ToolGlobals: CDFToolConfig, file: str, drop: bool, directory=None) -> None:
     """Load raw data from csv files into CDF Raw
 
     Args:
         file: name of file to load, if empty load all files
         drop: whether to drop existing data
     """
+    if directory is None:
+        directory = f"./examples/{ToolGlobals.example}/data/raw"
     client = ToolGlobals.verify_client(capabilities={"rawAcl": ["READ", "WRITE"]})
     # The name of the raw database to create is picked up from the inventory.py file, which
     # again is templated with cookiecutter based on the user's input.
@@ -74,7 +78,7 @@ def load_raw(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
         files.append(file)
     else:
         # Pick up all the .csv files in the data folder of the example.
-        for _, _, filenames in os.walk(f"./examples/{ToolGlobals.example}/data/raw"):
+        for _, _, filenames in os.walk(directory):
             for f in filenames:
                 if ".csv" in f:
                     files.append(f)
@@ -82,7 +86,7 @@ def load_raw(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
         return
     print(f"Uploading {len(files)} .csv files to {raw_db} RAW database...")
     for f in files:
-        with open(f"./examples/{ToolGlobals.example}/data/raw/{f}", "rt") as file:
+        with open(f"{directory}/{f}", "rt") as file:
             dataframe = pd.read_csv(file, dtype=str)
             dataframe = dataframe.fillna("")
             try:
@@ -100,7 +104,11 @@ def load_raw(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
     print(f"Successfully uploaded {len(files)} raw csv files to {raw_db} RAW database.")
 
 
-def load_files(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
+def load_files(
+    ToolGlobals: CDFToolConfig, file: str, drop: bool, directory=None
+) -> None:
+    if directory is None:
+        directory = f"./examples/{ToolGlobals.example}/data/files"
     try:
         client = ToolGlobals.verify_client(capabilities={"filesAcl": ["READ", "WRITE"]})
         files = []
@@ -108,9 +116,7 @@ def load_files(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
             files.append(file)
         else:
             # Pick up all the files in the files folder of the example.
-            for _, _, filenames in os.walk(
-                f"./examples/{ToolGlobals.example}/data/files"
-            ):
+            for _, _, filenames in os.walk(directory):
                 for f in filenames:
                     files.append(f)
         if len(files) == 0:
@@ -118,7 +124,7 @@ def load_files(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
         print(f"Uploading {len(files)} files/documents to CDF...")
         for f in files:
             client.files.upload(
-                path=f"./examples/{ToolGlobals.example}/data/files/{f}",
+                path=f"{directory}/{f}",
                 data_set_id=ToolGlobals.data_set_id,
                 name=f,
                 external_id=ToolGlobals.example + "_" + f,
@@ -134,12 +140,20 @@ def load_files(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
         return
 
 
-def load_timeseries(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
-    load_timeseries_metadata(ToolGlobals, file, drop)
-    load_timeseries_datapoints(ToolGlobals, file)
+def load_timeseries(
+    ToolGlobals: CDFToolConfig, file: str, drop: bool, directory=None
+) -> None:
+    load_timeseries_metadata(ToolGlobals, file, drop, directory=directory)
+    if directory is not None:
+        directory = f"{directory}/datapoints"
+    load_timeseries_datapoints(ToolGlobals, file, directory=directory)
 
 
-def load_timeseries_metadata(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
+def load_timeseries_metadata(
+    ToolGlobals: CDFToolConfig, file: str, drop: bool, directory=None
+) -> None:
+    if directory is None:
+        directory = f"./examples/{ToolGlobals.example}/data/timeseries"
     client = ToolGlobals.verify_client(
         capabilities={"timeseriesAcl": ["READ", "WRITE"]}
     )
@@ -149,21 +163,20 @@ def load_timeseries_metadata(ToolGlobals: CDFToolConfig, file: str, drop: bool) 
         files.append(file)
     else:
         # Pick up all the .json files in the data folder of the example.
-        for _, _, filenames in os.walk(
-            f"./examples/{ToolGlobals.example}/data/timeseries/"
-        ):
+        for _, _, filenames in os.walk(directory):
             for f in filenames:
                 if ".json" in f:
                     files.append(f)
     # Read timeseries metadata
     timeseries: list[TimeSeries] = []
     for f in files:
-        with open(
-            f"./examples/{ToolGlobals.example}/data/timeseries/{f}", "rt"
-        ) as file:
+        with open(f"{directory}/{f}", "rt") as file:
             ts = json.load(file)
             for t in ts:
-                timeseries.append(TimeSeries._load(t))
+                ts = TimeSeries()
+                for k, v in t.items():
+                    ts.__setattr__(k, v)
+                timeseries.append(ts)
     if len(timeseries) == 0:
         return
     drop_ts: list[str] = []
@@ -192,7 +205,11 @@ def load_timeseries_metadata(ToolGlobals: CDFToolConfig, file: str, drop: bool) 
     print(f"Loaded {len(timeseries)} timeseries from {len(files)} files.")
 
 
-def load_timeseries_datapoints(ToolGlobals: CDFToolConfig, file: str) -> None:
+def load_timeseries_datapoints(
+    ToolGlobals: CDFToolConfig, file: str, directory=None
+) -> None:
+    if directory is None:
+        directory = f"./examples/{ToolGlobals.example}/data/timeseries/datapoints"
     client = ToolGlobals.verify_client(
         capabilities={"timeseriesAcl": ["READ", "WRITE"]}
     )
@@ -202,9 +219,7 @@ def load_timeseries_datapoints(ToolGlobals: CDFToolConfig, file: str) -> None:
         files.append(file)
     else:
         # Pick up all the .csv files in the data folder of the example.
-        for _, _, filenames in os.walk(
-            f"./examples/{ToolGlobals.example}/data/timeseries/datapoints/"
-        ):
+        for _, _, filenames in os.walk(directory):
             for f in filenames:
                 if ".csv" in f:
                     files.append(f)
@@ -213,9 +228,7 @@ def load_timeseries_datapoints(ToolGlobals: CDFToolConfig, file: str) -> None:
     print(f"Uploading {len(files)} .csv file(s) as datapoints to CDF timeseries...")
     try:
         for f in files:
-            with open(
-                f"./examples/{ToolGlobals.example}/data/timeseries/datapoints/{f}", "rt"
-            ) as file:
+            with open(f"{directory}/{f}", "rt") as file:
                 dataframe = pd.read_csv(file, parse_dates=True, index_col=0)
             print(f"Uploading {f} as datapoints to CDF timeseries...")
             client.time_series.data.insert_dataframe(dataframe)
@@ -227,24 +240,24 @@ def load_timeseries_datapoints(ToolGlobals: CDFToolConfig, file: str) -> None:
         return
 
 
-def load_transformations(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> None:
+def load_transformations(
+    ToolGlobals: CDFToolConfig, file: str, drop: bool, directory=None
+) -> None:
+    if directory is None:
+        directory = f"./examples/{ToolGlobals.example}/transformations"
     client = ToolGlobals.verify_client(
         capabilities={"transformationsAcl": ["READ", "WRITE"]}
     )
     tmp = ""
     if file:
         # Only load the supplied filename.
-        os.mkdir(f"./examples/{ToolGlobals.example}/transformations/tmp")
-        os.system(
-            f"cp ./examples/{ToolGlobals.example}/transformations/{file} ./examples/{ToolGlobals.example}/transformations/tmp/"
-        )
+        os.mkdir(f"{directory}/tmp")
+        os.system(f"cp {directory}/{file} {directory}/tmp/")
         tmp = "tmp/"
-    configs = parse_transformation_configs(
-        f"./examples/{ToolGlobals.example}/transformations/" + tmp
-    )
+    configs = parse_transformation_configs(f"{directory}/{tmp}")
     if len(tmp) > 0:
-        os.system(f"rm -rf ./examples/{ToolGlobals.example}/transformations/tmp")
-    cluster = os.environ.get("CDF_CLUSTER", "westeurope-1")
+        os.system(f"rm -rf {directory}/tmp")
+    cluster = ToolGlobals.environ("CDF_CLUSTER")
     transformations = [
         to_transformation(client, conf_path, configs[conf_path], cluster)
         for conf_path in configs
@@ -279,3 +292,42 @@ def load_transformations(ToolGlobals: CDFToolConfig, file: str, drop: bool) -> N
     print(
         f"Created {len(created_transformations)} transformations for example {ToolGlobals.example}."
     )
+
+
+def load_readwrite_group(
+    ToolGlobals: CDFToolConfig,
+    capabilities: List[Dict[str, Any]],
+    source_id="readwrite",
+) -> None:
+    client = ToolGlobals.verify_client(
+        capabilities={"groupsAcl": ["LIST", "READ", "CREATE", "DELETE"]}
+    )
+    try:
+        groups = client.iam.groups.list(all=True)
+    except Exception as e:
+        print(f"Failed to retrieve groups.")
+        ToolGlobals.failed = True
+        return
+    old_group_id = None
+    for group in groups:
+        if group.source_id == source_id:
+            old_group_id = group.id
+            break
+    try:
+        group = client.iam.groups.create(
+            Group(
+                name=source_id,
+                source_id=source_id,
+                capabilities=capabilities,
+            )
+        )
+    except Exception as e:
+        print(f"Failed to create group {source_id}.")
+        ToolGlobals.failed = True
+        return
+    if old_group_id:
+        try:
+            client.iam.groups.delete(id=old_group_id)
+        except Exception as e:
+            print(f"Failed to delete group {old_group_id}.")
+            ToolGlobals.failed = True